rename api module to repository
