+++ core/repository/proguard-rules.pro
