+++ core/repository/src/main/AndroidManifest.xml
