+++ core/repository/src/main/java/jp/co/yumemi/repository/di/ApiModule.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.api.di
+package jp.co.yumemi.repository.di
 
 import org.koin.core.annotation.ComponentScan
 import org.koin.core.annotation.Module
