+++ core/repository/src/main/java/jp/co/yumemi/repository/model/WeatherResponse.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.api.model
+package jp.co.yumemi.repository.model
 
 import com.squareup.moshi.JsonClass
 import java.util.Date
